+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 
 * :bug:`-` Fix an issue where users cannot authenticate Gnosis Pay using Zeal Wallet.
 * :bug:`-` Fix an issue where the WalletConnect modal is not visible.
+* :bug:`11147` rotki will now properly detect staked GRT balances on Arbitrum after The Graph's Horizon upgrade.
 * :bug:`11149` rotki will now properly pull all new Coinbase events.
 * :bug:`-` Asset selection fields will now properly display ignored assets if they are already selected as the value.
 * :bug:`11113` An invalid Coinbase API key in the DB will no longer prevent logging into the app.